+++ contracts/USDCclaimToken.sol
@@ -4,7 +4,7 @@ pragma solidity 0.8.10;
 import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 
 //--------------------------------------
-//   USDC claim Token Contract 
+//   USDC claim Token Contract
 //
 // Symbol      : USDCCT
 // Name        : USDCClaimToken
@@ -19,7 +19,7 @@ contract USDCclaimToken is IERC20 {
     uint256 public initialSupply;
     string public name;
     string public symbol;
-    
+
     mapping(address => uint256) internal balances;
     mapping(address => mapping(address => uint256)) internal allowed;
     mapping(address => bool) internal admins;
@@ -34,17 +34,16 @@ contract USDCclaimToken is IERC20 {
         _;
     }
 
-    constructor()  {
+    constructor() {
         name = "LFUSDCLP";
         symbol = "LFUSDCLP";
         decimals = 18;
         _totalSupply = 0;
-	    initialSupply = _totalSupply;
+        initialSupply = _totalSupply;
         owner = msg.sender;
-	    balances[msg.sender] = _totalSupply;
+        balances[msg.sender] = _totalSupply;
         emit Transfer(address(0), msg.sender, _totalSupply);
     }
-
 
     function addAdmin(address account) public onlyOwner {
         admins[account] = true;
@@ -60,55 +59,89 @@ contract USDCclaimToken is IERC20 {
 
     function totalSupply() external view override returns (uint256) {
         return _totalSupply - balances[address(0)];
-        
     }
 
-    function balanceOf(address tokenOwner) external view override returns (uint getBalance) {
+    function balanceOf(address tokenOwner)
+        external
+        view
+        override
+        returns (uint256 getBalance)
+    {
         return balances[tokenOwner];
     }
 
-    function allowance(address tokenOwner, address spender) external view override returns (uint remaining) {
+    function allowance(address tokenOwner, address spender)
+        external
+        view
+        override
+        returns (uint256 remaining)
+    {
         return allowed[tokenOwner][spender];
     }
- 
-    function approve(address spender, uint tokens) external override returns (bool success) {
+
+    function approve(address spender, uint256 tokens)
+        external
+        override
+        returns (bool success)
+    {
         allowed[msg.sender][spender] = tokens;
         emit Approval(msg.sender, spender, tokens);
         return true;
     }
-    
-    function transfer(address to, uint tokens) external override returns (bool success) {
-        require(to != address(0),"USDCclaimToken: Address should not be a zero");
+
+    function transfer(address to, uint256 tokens)
+        external
+        override
+        returns (bool success)
+    {
+        require(
+            to != address(0),
+            "USDCclaimToken: Address should not be a zero"
+        );
         balances[msg.sender] = balances[msg.sender] - tokens;
         balances[to] = balances[to] + tokens;
         emit Transfer(msg.sender, to, tokens);
         return true;
     }
-    
-    function transferFrom(address from, address to, uint tokens) external override returns (bool success) {
-        require(to != address(0),"USDCclaimToken: Address should not be a zero");
+
+    function transferFrom(
+        address from,
+        address to,
+        uint256 tokens
+    ) external override returns (bool success) {
+        require(
+            to != address(0),
+            "USDCclaimToken: Address should not be a zero"
+        );
         balances[from] = balances[from] - tokens;
         allowed[from][msg.sender] = allowed[from][msg.sender] - tokens;
         balances[to] = balances[to] + tokens;
         emit Transfer(from, to, tokens);
         return true;
     }
-   
-    function burn(address account,uint tokens) external onlyAdmin {
-        require(account != address(0),"USDCclaimToken: Burn from a zero address");
+
+    function burn(address account, uint256 tokens) external onlyAdmin {
+        require(
+            account != address(0),
+            "USDCclaimToken: Burn from a zero address"
+        );
         uint256 accountBalance = balances[account];
-        require(accountBalance >= tokens , "USDCclaimToken: Burn amount exceeds Balance");
+        require(
+            accountBalance >= tokens,
+            "USDCclaimToken: Burn amount exceeds Balance"
+        );
         balances[account] = accountBalance - tokens;
         _totalSupply = _totalSupply - tokens;
-        emit Transfer(account,address(0), tokens);
+        emit Transfer(account, address(0), tokens);
     }
-    
-    function mint(address account,uint tokens) external onlyAdmin {
-        require(account != address(0),"USDCclaimToken: Mint from a zero address");
+
+    function mint(address account, uint256 tokens) external onlyAdmin {
+        require(
+            account != address(0),
+            "USDCclaimToken: Mint from a zero address"
+        );
         balances[account] = balances[account] + tokens;
         _totalSupply = _totalSupply + tokens;
-        emit Transfer(address(0),account,tokens); 
+        emit Transfer(address(0), account, tokens);
     }
-
-
 }
